+++ client/src/components/TerrainTiles.tsx
@@ -1,12 +1,10 @@
-import styled from '@emotion/styled';
 import { observer } from 'mobx-react';
-import { Position, TileTable } from 'core/schemas';
+import { TileTable } from 'core/schemas';
 import { worldState } from '..';
 import Marsh from 'assets/Marsh.png';
 import Bush from 'assets/Bush.png';
 import Grass from 'assets/Grass.png';
-
-export const tileSideWidth = 70;
+import { PositionWrapper } from 'components/Other';
 
 const TerrainTile = observer(() => {
   const tiles = TileTable.getAll(worldState.tableState);
@@ -35,21 +33,5 @@ const TerrainTile = observer(() => {
     </div>
   );
 });
-
-interface PositionWrapperProps {
-  position: Position;
-}
-
-export const PositionWrapper = styled.div<PositionWrapperProps>`
-  position: absolute;
-  left: ${(props) => props.position.x * tileSideWidth}px;
-  bottom: ${(props) => props.position.y * tileSideWidth}px;
-  width: ${tileSideWidth}px;
-  height: ${tileSideWidth}px;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-`;
 
 export default TerrainTile;
