+++ client/src/index.tsx
@@ -1,8 +1,8 @@
 import ReactDOM from 'react-dom/client';
 import { AppRouter } from './pages/AppRouter';
 import { ChakraProvider } from '@chakra-ui/react';
-import { WorldState } from './store/stateStore';
-import { UIState } from './store/uiStore';
+import { WorldState } from './keystone/stateStore';
+import { UIState } from './core/uiStore';
 import { makeAutoObservable } from 'mobx';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
