+++ client/src/keystone/middleware.ts
@@ -1,0 +1,14 @@
+export interface KeystoneTx<T> {
+  Headers: { [p: string]: any };
+  Data: T;
+}
+
+export type HeaderEntry<T> = [string, T];
+
+export function NewKeystoneTx<T>(request: T, ...headerEntries: HeaderEntry<any>[]): KeystoneTx<T> {
+  const headersMapJSON = Object.fromEntries(headerEntries);
+  return {
+    Headers: headersMapJSON,
+    Data: request,
+  };
+}
